+++ src/trainer.h
@@ -4,14 +4,14 @@
 #include "types.h"
 #include "util.h"
 
-float TotalError(DataSet* data, NN* nn);
+double TotalError(DataSet* data, NN* nn);
 void Train(int batch, DataSet* data, NN* nn, NNGradients* g, BatchGradients* local);
 
-INLINE float WDL(int8_t wdl) { return wdl / 2.0f; }
+INLINE double WDL(int8_t wdl) { return wdl / 2.0; }
 
-INLINE float Error(float r, DataEntry* e) {
-  return 0.5f * powf(r - WDL(e->wdl), 2.0f) + 0.5f * powf(r - e->eval, 2.0f);
+INLINE double Error(double r, DataEntry* e) {
+  return 0.5 * pow(r - WDL(e->wdl), 2.0) + 0.5 * pow(r - e->eval, 2.0);
 }
-INLINE float ErrorGradient(float r, DataEntry* e) { return (r - WDL(e->wdl)) + (r - e->eval); }
+INLINE double ErrorGradient(double r, DataEntry* e) { return (r - WDL(e->wdl)) + (r - e->eval); }
 
 #endif
